osutil: use tilde suffix for temporary files used for atomic replacement
This patch changes the temporary files used for, amongst other, apparmor
profiles with ~ (tilde) so that they are ignored by apparmor tools in an
unlikely but real case when snapd is killed after creating the temporary
file but before removing it or renaming it over to something else.

Tilde is the best suffix for this choice as other suffixes that are
ignored include packaging specific semantics (like .dpkg-new) or have
well-known meaning (.rej, .orig).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>