cmd/snap-update-ns: allow path guardian to inspect fs ops
This patch adds the concept of a path guardian which can check for
trespassing ahead of a filesystem write operation. The guardian is
actually not implemented in this patch, merely the infrastructure is
laid out so that the guardian can operate later on.

This patch is a long, no-op, that will allow the subsequent patch to
meaningfully improve the system.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>