overlord/ifacestate: store and use revision with security profiles set
When snapd restarts while a snap is being refreshed we can get into a
state when a snap is installed but all revisions are inactive because we
are mid-way through the refresh process. When that happens the
interfaces of said snap were in the interface repository, ready to be
used (connected, set up, or otherwise). If snapd restarts at that time
that fact is lost because only active snaps are added to the repository
on startup.

This patch adds a new map to the state, to store the revision for which
we have setup the snap's security profiles. This can be used in certain
situations (do-undo logic) without jumping through hoops.

The revision is set and reset when the snap is successfully set up or
removed by the interface manager.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>