many: implement a poor man's privileges drop, use for auth.json
This PR conceptually implements support for allowing you to do i/o on
a file as a less privileged user than you're currently running.

The nice, in-process approach doesn't work, as it needs Go 1.10's
semantics around runtime.LockOSThread.

This instead uses a helper, `privhelper`, that lives in libexec, that
knows how to read, write and remove a file, changing to a certain user
ahead of time.

We can't just use `su`, because su has strange requirements around
running from a terminal, and the test suite barfs on it (in practice
it might work acceptably, but it being untestable is a big deal).

We'd rather not use `sudo`, as this is for something that might
already be running under `sudo` and nesting sudos confuses some
things; better stay clear.

So this PR implements `sys.Setuid` and `sys.Setgid`, with the same
caveats as they had in Go 1.3: it only affects the current thread (so
LockOSThread rears its head again).

This is the first helper that's meant to be run from `snap`, so
`cmd.InternalToolPath` would panic at it. Perhaps a second iteration
would rework that so that the current approach of having logic in
`dirs` is less weird.