packaging/amzn2: initial packaging of 2.32.3 for Amazon Linux 2
Initial packaging of snapd for Amazon Linux 2. The spec is based mostly on
Fedora spec file, with minor pieces picked from openSUSE one.

Note, that the snaps are mounted under /var/lib/snapd/snap. The /snap
directory (or symlink) is not provided by the package, so classic snaps do not
work out of the box.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>