many: fix "refresh-mode: sig{term,hup,usr[12]}" via KillMode=process
When the main process of a unit gets killed systemd will also kill
all the other processes in the cgroup. So the current approach of
using "systemctl kill --kill-who=main snap-service" is not sufficient
to keep the child processes alive. We need to add "KillMode=process"
to the unit. The side effect of this is that a normal stop also no
longer stops all processes in the cgroup. AFAICT there is no option
to have both (kill only kills main, stop stops all).