many: fix "refresh-mode: sig{term,hup,usr[12]}" via KillMode=process
When the main process of a unit gets killed systemd notices that
(it watches for sigchld) and the unit will enter the stop state.
This means that systemd will do the usual cleanup when a unit stops
which may mean killing all processes in the cgroup of the unit.

This means the current approach of using
"systemctl kill --kill-who=main snap-service" is not sufficient
to keep the child processes alive.

We need to add "KillMode=process" to the unit so that systemd does
not kill the remaining processes in the cgroup. The side effect of
this is that a normal `systemctl stop snap...` now also no longer
stops all processes in the cgroup. To mitigate this we explicitly
kill all processes of the units when removing the snap.