interfaces/apparmor: workaround kernel memory leak
This patch changes the apparmor backend to insert a randomized rule when
running on classic devices. Core devices are mostly on a kernel that is
not affected (4.4 and similar). Classic devices run various kernels,
specifically 4.13 and newer. Since 4.13 the kernel leaks a significant
amount of memory when an apparmor profile is loaded (with the intent to
replace an existing profile) but that profile is identical to the one
currently loaded.

To avoid this all apparmor profiles now contain a randomized component.
The component is set to zero on an unaffected kernel and to a random
integer otherwise. The integer is subsequently used to create an
apparmor profile rule for the path "/.bug/lp1750594/workaround.$RANDOM",
granting read permission.

Related-To: https://bugs.launchpad.net/apparmor/+bug/1750594
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>