tests: run interfaces-boradcom-asic-control early
We need to run the broadcom-asic-control test very early. The reason is
that this test will modprobe linux-*-bde which will try to allocate a
huge page. One the system ran for a while no chunk of memory like this
will be not be available and the kernel will oops with:

  "modprobe: page allocation failure: order:7, mode:0xxxx"