cmd/snap-update-ns: rework trespassing on /
When we are trespassing on a directory we are really trespassing on the first
thing that is not a tmpfs or a squashfs (typically a regular filesystem like
ext4) that we encounter as we descend from /.

On classic systems, the directory we are technically trespassing on is / itself
is / is pretty much always writable there. This creates an issue where we would
be supposed to create a writable mimic at / to "conceal" the changes from the
outer mount namespace.

Because we cannot construct mimics at / we must instead construct them at the
next best place, that is the immediately following directory node, such as
/etc.

The code that was doing that was somewhat magic so I re-structured it to be
more explicit about what is going on and added some comments to explain the
rationale.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>