snap,tests : don't fail if we cannot stat MountFile
The snap mount file may be a dangling symlink in case of try-mode snaps
that got moved around by the user. In those cases the snap is still
mounted so it doesn't show as broken but it cannot be stat'ed. Such
snaps will now show as broken as the returned error NotFoundError has
special meaning elsewhere in the code

Fixes: https://bugs.launchpad.net/snapd/+bug/1764977
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>