snap,overlord/snapstate: make BrokenSnapError an interface
There are now many (now private) errors that can be handled as a
BrokenSnapError. The existing NotFoundError becomes notFoundError and is
joined by corrputMetaError. The latter error encompasses issues with
incorrect or invalid YAML, missing hooks or other things involving the
meta directory.

The existing code that used the type is adjusted to use the new
interface.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>