interfaces/apparmor: allow writable mimic in parent directories
This patch fixes the layout system to allow creation of a writable mimic
in any reasonable parent of the affected layout source or target path.

When the layout for /usr/lib/foo/bar/ is requested and /usr/lib/foo/
doesn't exist we really need a rule that allows any path of
/usr/lib/foo/bar/ not to exist.

To illustrate the issue, let's say that /var/snap/mysnap/123/a/b/c/
is laid out on top of /var/cache/mysnap/a/b/c/. Normally we would allow
a writable mimic to be constructed at /var/cache/mysnap/share/a/b/ so that
"c" may be created. As is evident by this trivial example it is unlikely
that /var/cache/mysnap exists and it thus it also may have to be created.

This is achieved with the "chop tree" of apparmor alternation expressions
with a fixed-size "trunk" of non-optional elements. In this case it allows
the mimic to be created in /var/{,cache/{,mysnap/,{a/{,b/}}}} since
we only assume /var exists.

This makes it possible to express natural layouts without understanding
the implementation or its limitations. The same patch adjusts the
content interface that uses some of the same helper functions.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>