snap: improve error for snaps not available in the given context
When we have a snap in the store that is not available in the
channel that the user requested we currently show a "not-found"
error. But it is better to display more meaningful information.

The store provides a more meaningful error code now, so we use
it. This is a resurrect from #4443.

Fixes https://bugs.launchpad.net/snapstore/+bug/1665787