New utility to save and restore the snap state
This reduces the time needed to restore the snapd state by the 50% of
the time needed aprox. For each restore this way is 10 seconds faster in
a pi2 or pi3 or dragonboard.

This change is reducing in 30 minutes the whole execution in the pi2.
old version: https://paste.ubuntu.com/p/vMYGKbjJJr/
new version: https://paste.ubuntu.com/p/Dh4mXrJdXj/