overlord,systemd: store snap revision in mount units
This patch changes the description of the systemd mount unit so that it
now refers to both the snap name and the snap revision.

When a snap refreshes the last three revisions seen are kept on disk.
When the system is shutting down or whenever someone is interacting with
systemd the identical descriptions of said mount units may be confusing.
To avoid the confusion the description now reads: "Mount unit for
snap-name, revision 123".

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>