dirs: on opensuse store apparmor profiles in /etc/apparmor.d
This is a RFC patch for moving apparmor profiles from
/var/lib/snapd/apparmor/profiles to /etc/apparmor.d
when running on openSUSE and LIKE distributions.

The main motivation is that /var/lib/snapd/apparmor/profiles is not
supported by the apparmor init scripts on openSUSE. The situation is
that all the distributions essentially have their own stack of apparmor
init code, different from one another. A massive rewrite is ongoing that
will, amongst others, allow to easily say that snapd stores apparmor
profiles in a different directory but this is still work in progress
upstream and nowhere to be seen downstream.

In an effort to improve the status of apparmor support on openSUSE I
would propose that we merge this change. It was acked by the apparmor
maintained in openSUSE (cboltz) and is something we could deprecate and
phase out over time, as the improved init script for apparmor becomes
commonplace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>