snap: do not use overly short timeout in `snap {start,stop,restart}`
* snap: do not use overly short timeout in `snap {start,stop,restart}`

Right now we are using a timeout of 5s for `snap start/stop/restart`.
For many snaps this is too short so this PR increases the limit to
61s. This improves the situation while ensuring that snaps cannot
hold a hook "hostage" for a long time.

Longer term we need I think we to refactor servicestate.go to
reuse the existing systemd code in the `systemd` and `wrappers`
packages which handle stop-timeout correctly already. We just
need to be careful that we enforce sensible limits (and disucss
what "sensible" means).