Fix interfaces-network test for systems with partial confinement
Currently this test is passing on debian but because there is a
connection error when tries to connect the snap to the service.

The test was modified to show the los when it works (shared in the next
link): https://paste.ubuntu.com/p/MHPCWjRMsv/

With the previous implementation the connection to the service worked
even if the interface was disconnected.