snapd.core-fixup.sh: add workaround for corrupted uboot.env
We see sometimes corrupted uboot.env files created by fsck.vfat.

On the fat filesystem they are indistinguishable because one
has a fat16 name UBOOT.ENV (and not lfn (long-file-name)) but
the other has a "uboot.env" lfn name and a FSCK0000.000 FAT16
name. The only known workaround is to remove all dupes and put
one file back in place.

Ideally we also fix fsck.vfat (possible fix in
https://github.com/dosfstools/dosfstools/pull/83)

With this patch the sequence of events will be:
1. snapd refreshes
2. snapd writes new uboot.env with snap_mode=try
3. snapd reboots
4. uboot starts loads the uboot.env and writes it with snap_mode=trying
5. system boots, /boot/uboot is set to fs_passno=2 and systemd checks
   it before it mounts it
6. the file is corrupted, most likely by the uboot write
7. the fsck in (6) renames the file to fsck0000.000 but leaves the
   l-f-n entries intact and now we have two files.
8. snapd.core-fixup.sh runs, detects the two files and removes them
   and just keeps one. In the (manual) tests I ran this was the
   file with "snap_mode=try"
9. snapd will assume that the boot failed because snap_core is still
   at the previous core and reverts
10. uboot will on revert write snap_mode=trying to the right file
    now and our failover logic works again.