many: add "snap debug sandbox-features" and needed bits (#5142)
* many: add "snap debug sandbox" and needed bits

This patch adds a way to inspect the features of the sandbox from a new
hidden, debugging command. The command shows a set of tags for each of
the security backends that are in use.

Having access to this command helps in two ways:

First of all, this is far more comprehensive and informative than the
very dry "partial" vs "full" output of "snap debug confinement". Snapd
offers a wide range of confinement technologies when used on a mainline
kernel and this should be celebrated.

Second of all having this allows us to write more precise tests. If a
test needs to be skipped because a specific thing is absent can now be
codified directly. This will allow both Debian and openSUSE to run far
more confinement-specific tests than we currently do simply because
"partial" apparmor is reported. Mainline kernel is very close to full
apparmor patch used by Ubuntu and the remaining differences don't apply
to many existing tests.

Technically this patch adds the key "sandbox", defined as a map of lists
of strings, to the sysinfo API and uses it in the new hidden command.
Each of the security backends can now optionally offer sandbox tags that
are associated with the backend name in the said map.

* interfaces: namespace kernel features with "kernel:"

This arrangement will allow the seccomp and apparmor backends to report
additional sandbox feature tags without risking any clash with kernel
reported things.

* interfaces/seccomp: return at least one sandbox tag from seccomp

In case the kernel doesn't advertise seccomp kernel features we don't
want to make the seccomp backend appear as missing or unused. Add an
explicitly managed tag to achieve that.

* interfaces/apparmor: dbus mediates more than just bus names

* many: rename sandbox tags to sandbox features

* cmd/snap: add support for "sandbox-features --required=..."

This patch allows using the sandbox-features command for simple
scripting, especially in spread tests, where we could check if a given
sandbox feature is present before checking confinement is effective.

* tests: fix and expand sandbox-features tests

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>