snapstate: make TestDoPrereqRetryWhenBaseInFlight less brittle
The TestDoPrereqRetryWhenBaseInFlight was failing in the autopkgtest
environment. It looks like the reason is that the VMs that run the
test are very slow (due to load) and this results in the timeouts
to miss. Rework the test to give it more time and also to be more
uniform with the previous bits of the test.