interfaces/process-control: only allow sending to other processes
The process-control interface previously used a bare signal rule which in
addition to allowing sending to all processes also allowed receiving from all
processes. Refine this rule to simply 'signal (send)' to make snaps plugging
the process-control to behave the same as other snaps with regard to receiving
signals (ie, we can still receive from unconfined and our snap's processes but
not from random non-snap-but-still-confined processes).