interfaces/builtin: initial version of the anbox-support interface
This implements the anbox-support interface as a copy of the existing
lxd-support interface. Anbox consists mainly of two components: A
container manager (similar to LXD) and a so called session manager. The
container manager uses LXC to spawn a container for the Android
operating system and will used a AppArmor profile derived from the one
LXD uses for its containers to confine it.

The interface basically allows the container manager to escape
confinement.