daemon: fix unit tests on arch
On Arch (or other distros where snap mount point is not /snap), the system-info
test fails as 'classic' is not returned in the list of confinement options. The
reason is that dirs.SetRootDir() comes before release.MockReleaseInfo() call and
we end up with host specific snap mount dir path. Make sure to reload the root
directory path before querying the API to ensure a stable output.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>