spread-shellcheck: convert to python
The original implementation used a inefficient pipeline (yq -> jq -> shellcheck)
and spawned a large number of processes. As a result, the command:

  NO_FAIL=1 MUST_PASS=tests/unit/spread-shellcheck/must V=1 ./spread-shellcheck \
      spread.yaml tests

Took a bit more than 3 minutes to run.

The patch rewrites spread-shellcheck in Python to make it more affordable to run
during the regular development work. The implementation only depends on PyYAML.
The run time of the same command is now ~30s.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>