interfaces/apparmor: use helper to load stray profile
A long time ago I ported all ad-hoc calls to apparmor_parser to go
through a helper function. At the time the location of the cache was
not something one could pass to the function so one of the original call
sites remained. Over time this limitation was lifted but that stray call
site was not updated to use the new function. This patch fixes it.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>