boot: support snap.TypeBase for booting
In the past boot/kernel_os.go:SetNextBoot() would only look at the
type: OS/Kernel to determine what to boot next. However with the
addition of bootable bases we can now also boot from type base.

In order to know which base to use for booting the bootloader code
now inspects the snap name of the last good boot and if those do
not match it will just ignore the snap.