selftest: add new selftest package that tests squashfs mounting
We got a number of bugreports that contain messages like:
```
5月 29 10:14:18 x4-640 systemd[1]: snap-core-4486.mount: Failed with result 'exit-code'.
5月 29 10:14:18 x4-640 systemd[1]: Failed to mount Mount unit for core.
```
which indicates that the system reporting the error is not capable
of mounting squashfs images (not squashfs kernel module or not with
xz compression).

Because squashfs mounting is essential for snapd to work we now test
this very early and stop the daemon if squashfs files cannot be
mounted.