snap: introduce new fields for parallel snap installation
Parallel installation of snaps allows a snap to be installed multiple times,
with instance carrying a unique local key appended to the the snap's store name.
Eg. foo is a locally installed snap, foo_bar is another instance of the snap
with local key 'bar'. The local key does not appear anywhere inside snap.yaml,
and is only assigned locally.

The patch introduces changes to basic snap data structures.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>