many: use ~/.config/snap if it exists, instead of ~/.snap
Before this change we were storing auth.json and the asserts keyring in
~/.snap.

This ignores the xdg base directory standard, which says that a user's
config for an app should go in ~/.config, and its data in
~/.local/share/ (and cache in ~/.cache), all overridable by the
appropriate env vars.

Whether auth.json and the asserts keyring is config or data is up for
debate; I've gone with it being config. Sadly neither ssh nor gnupg
themselves follow the standard so I wasn't sure where to look. Config
and data are often hard to separate.

This change comes up when thinking through the warnings subsystem,
which requires a client-side nonce.