snap,client: Show "publisher" in `snap list` and expose in client API
In the client package we have a "Developer" entry in the Snap
struct. However this name is actually misleading as this is
set to the publisher of the snap. To decrease confusion this
PR adds a "Publisher" field that is also set to the publisher.

This means that with PR#5234 (`snap list --format=help`) we
can now show a meaningful field.

This PR also modifies the output of `snap list` to show
"publisher" in the header instead of "developer" to reflect
reality.