many: rename snap.Info methods *Name() and *Dir(), add helpers, prep for parallel installs
Rename snap.Info methods, namely:
 - Name() -> InstanceName()
 - *Dir() -> Instance*Dir()

This triggers all the places where those methods are used to be updated and
reconsidered.

Introduced a number of helpers in snap for building various directory paths. The
helpers explicitly take snap name, so are fine to be use with both instance and
store snap names.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>