cmd/snap-update-ns: add IsReadOnly and tests
This patch adds a simple helper that checks if a directory is read only
in a way that sandboxed processes cannot overcome. This will be used in
upcoming trespassing checks where attempts to write to read only
directories are honored as they trigger EROFS which is handled by the
construction of a writable mimic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>