errtracker: do not send duplicated reports (#5263)
* errtracker: do not send duplicated reports

This will add a 7 day history for the error tracker. It will not
send the same report over and over again but instead skip reports
that were already send.

Reviews asked for locking and this means we need something better
than the hand implmented DB. Switch to boltdb because we already
use it for the advisor interface.