interfaces/apparmor: rewrite chopPath to avoid apparmor bugs
This patch rewrites chopPath to avoid a bug in the apparmor_parser that
manifests itself when alternation patterns are used with directories and
wildcards. For more details see
https://bugs.launchpad.net/apparmor/+bug/1769971

The idea is to simply return a pair of lists of path expressions that
are longer but equivalent to the older path alternations.  The code is
also simplified and shortened thanks to the use of the PathIterator.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>