put snap-confine profile from core/snapd into /var/lib/snapd/apparmor/profiles
Previously we put the auto-generated profile for snap-confine from
the core snap into /etc/apparmor.d. With the need to run snap-confine
from the snapd snap on core18 systems this is no longer feasible
because /etc/apparmor.d is read-only (that is a desirable feature we
want to keep).

For this reason put the profile into /var/lib/snapd/apparmor/profiles
under the "snap-confine.*" prefix. The prefix means that no snaps
can interfere with this profile. It also means the profile gets
loaded on system startup just like any other snap profile.