Show executed tests on current system on failed tests
The idea of this change is to have information about which tests were
executed before a failed test and in which order. This will help to
reproduce errors much easier when the test failes because another does
not clean the environment as it should. This is usefull on when many
workers are used and it is not clear which tests were executed before.