Fix the delayed download of the big snap
After a research it was determined that the download is taking
somethimes more than 20 seconds to start (the GET does not finishes).

See this log: https://paste.ubuntu.com/p/9C5FHbx23d/

Also we can see that the "kill $(pidof snap)" does not work when for
example the process snap userd is alive. This produces a problem because
the snap download will be still alive and the snap is downloaded after
the test fails.

See PR #5329 for more information about the debug lines introduced.

In the following log we can see that the GET done through the line
"s.doRequest(ctx, httputil.NewHTTPClient(nil), reqOptions, user)" in the
file store/store.go does not finishes during the 20 seconds which we are
checking the partial file size.

log: https://paste.ubuntu.com/p/tbdcGfBshS/

So, this change is adding more time to the retry checking the partial
file size, and changes the kill command used to finish the snap
download.