interfaces: treat "snapd" snap as type:os
This patch changes the policy checker to internally detect the snap with
IDs of the "snapd" (in either staging or production) and translate its
type to "core". This allows the base policy to remain unchanged while
implicit interfaces are moved to "snapd" snap.

This also includes snap-id checking improvements from mvo (thanks)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>