tests: disable auto-refresh test on core18
Disable auto-refresh on core18 until we have a Canonical model
assertion here.

core18 right now does not have a canonical model. This means
that it will not get a serial. We code in devicestate.go that will only
try to auto-refresh after the systems has tried at least 3 times to get
a serial. But this means this test will timeout because the first retry
for the serial happens after 5min, then 10min, then 20min.